refactor(inference): extract response sanitization into sanitize.go
Pure code move, no behavior change. chatbot.go had grown to ~1225 lines; the
client-facing response-sanitization cluster (leak-field stripping #184, the
openrouter.reasoning redacted_thinking drop #373, reasoning/thinking token clamp
#374, body decompression, and the per-SSE-line helper) is a cohesive unit and
now lives in sanitize.go alongside the pre-existing sanitize_test.go. chatbot.go
drops to ~832 lines and keeps the request/response orchestration + billing.

Moved: leakKeys* vars, stripLeakKeys, isZeroNumber, openrouterLeakPrefix,
isLeakedRedactedThinking, stripUpstreamReasoningBlocks, clampReasoningTokenDetails,
clampTokenDetail, jsonNumberToInt, sanitizeResponseBody, isCompressedEncoding,
decodeBody, sanitizeStreamLine. compress/zlib and strconv imports move with them
(no longer used in chatbot.go; initializeReader still uses gzip/flate/brotli).

build + vet + full ctrl test suite green.
